third_party/blink/web_tests/external/wpt/fetch/range/general.window-expected.txt

This is a testharness.js-based test.
PASS Script executed from partial response
FAIL Fetch with range header will be sent with Accept-Encoding: identity assert_equals: Expect identity accept-encoding on media request expected "identity" but got "identity;q=1, *;q=0"
Harness: the test ran to completion.
